stock/backend/requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
PyJWT==2.8.0
schedule==1.2.0
Werkzeug==2.3.7
flask-cors==4.0.0 